esp01: Change serial to 74880 baud to match bootloader
The ESP-01 is supposed to come with a 40 MHz crystal, which would result
in 115200 baud for the bootloader. However, most ESP-01 modules come
with a 26 MHz crystal instead, resulting in this unusual boot
configuration of 74800 baud, which is 115200*26/40.
To avoid the garbage output from the bootloader in the serial monitor on
reset, use the same setting in the applications.

https://arduino-esp8266.readthedocs.io/en/3.0.2/ideoptions.html#crystal-frequency